apps/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "essentia-tensorflow>=2.1b6.dev1177",
    "fastapi[standard]>=0.116.1",
    "pydantic>=2.11.7",
    "yt-dlp>=2025.6.30",
    "numpy<2.0",
]

[project.scripts]
serve = "uvicorn main:app --reload"
